requirements.txt
Django>=5.1.7
requests>=2.0.0
python-dotenv>=1.0.0
